+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.0.2
+
+- Fix colors not displaying in flutter
+- Add success level
+
 ## 1.0.1
 
 - Readme update